+++ R/pin_default.R
@@ -21,5 +21,17 @@ pin_preview.default <- function(x, board = NULL, ...) {
 #' @keywords internal
 #' @export
 pin_load.default <- function(path, ...) {
-  readRDS(file.path(path, "data.rds"))
+  result <- readRDS(file.path(path, "data.rds"))
+
+  if ("AsIs" %in% class(result)) {
+    class(result) <- class(result)[class(result) != "AsIs"]
+  }
+
+  result
+}
+
+#' @keywords internal
+#' @export
+pin_fetch.default <- function(path, ...) {
+  path
 }