+++ R/utils.R
@@ -7,6 +7,9 @@ get_function <- function(name, package) {
   }
 }
 
-pins_show_progress <- function() {
-  identical(getOption("pins.progress", FALSE), TRUE)
+pins_show_progress <- function(size = 0) {
+  if (is.character(size)) size <- as.integer(size)
+
+  large_file <- getOption("pins.progress.size", 10^7)
+  identical(getOption("pins.progress", size > large_file), TRUE)
 }