+++ inst/doc/boards-websites.R
@@ -5,5 +5,11 @@
 #  pin_find(board = "txtexample")
 
 ## ----eval=FALSE----------------------------------------------------------
+#  pin_find(board = "txtexample", extended = TRUE)
+
+## ----eval=FALSE----------------------------------------------------------
+#  pin_info("mtcars", board = "txtexample")
+
+## ----eval=FALSE----------------------------------------------------------
 #  pin_get("mtcars", board = "txtexample")